requirements.txt
aiogram>=3.0
aiosqlite
openai
python-dotenv
fpdf2
